requirements.txt
numpy==1.26.4
pillow==10.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyserial==3.5
